# Lokata.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lokata.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Umpire": {
      "get": {
        "tags": [
          "Umpire"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Umpire"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Umpire"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Umpire"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Umpire"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Umpire/{id}": {
      "get": {
        "tags": [
          "Umpire"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Umpire"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Umpire"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Umpire"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Umpire"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Umpire"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Umpire"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "postalPlace": {
            "type": "string",
            "nullable": true
          },
          "places": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Approach": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitionsId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "instructorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "umpireId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competition": {
            "$ref": "#/components/schemas/Competition"
          },
          "competitions": {
            "$ref": "#/components/schemas/Competitions"
          },
          "instructor": {
            "$ref": "#/components/schemas/Instructor"
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          },
          "umpire": {
            "$ref": "#/components/schemas/Umpire"
          }
        },
        "additionalProperties": false
      },
...
      "Competition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "seriesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPointsPerSeries": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Competitions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "additionalProperties": false
      },
      "Competitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDisabledPerson": {
            "type": "boolean",
            "nullable": true
          },
          "professional": {
            "type": "boolean",
            "nullable": true
          },
          "sexId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "fileContent": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Instructor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "documentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "document": {
            "$ref": "#/components/schemas/Document"
          },
          "documentValidUntil": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shootingPlacesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "addressNavigation": {
            "$ref": "#/components/schemas/Address"
          },
          "competitionsList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competitions"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "Sex": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "asMale": {
            "type": "boolean"
          },
          "asFemale": {
            "type": "boolean"
          },
          "competitorList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competitor"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TakePlace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tookPlace": {
            "type": "boolean",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "TargetsAndCardsPhoto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "TargetsOrCardTake": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cardOrTargetTaken": {
            "type": "boolean",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "Umpire": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "permissionDocumentNumber": {
            "type": "string",
            "nullable": true
          },
          "permissionDocumentDate": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "permissionValidUntil": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}